requirements.txt
# Core Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Deep Learning Frameworks
torch==2.1.2
torchvision==0.16.2
tensorflow==2.15.0

# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
mediapipe==0.10.9
dlib==19.24.2
face-recognition==1.3.0

# Image Processing
Pillow==10.2.0
scikit-image==0.22.0
albumentations==1.3.1

# Data Science & ML
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Explainability
grad-cam==1.4.8
lime==0.2.0.1
shap==0.44.0

# Video Processing
moviepy==1.0.3
imageio==2.33.1
imageio-ffmpeg==0.4.9

# Model Training & Optimization
tensorboard==2.15.1
tqdm==4.66.1
timm==0.9.12  # PyTorch Image Models

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
jsonschema==4.20.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0

# Deployment
gunicorn==21.2.0
waitress==2.1.2

# Security
cryptography==41.0.7
python-jose==3.3.0

# File Handling
python-magic==0.4.27

# API Documentation
flask-swagger-ui==4.11.1

# Async Support
celery==5.3.4
redis==5.0.1
